// src/front/src/api/identity/models/appRole.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export type AppRole = (typeof AppRole)[keyof typeof AppRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AppRole = {
	Root: "Root",
	Ops: "Ops",
	Admin: "Admin",
	Support: "Support",
	User: "User",
} as const;